Parse Or patterns
